moved around files into a more logic file structure
