+++ package.json
@@ -2,8 +2,8 @@
   "name": "pino-http",
   "version": "7.0.0",
   "description": "High-speed HTTP logger for Node.js",
-  "main": "logger.js",
-  "types": "index.d.ts",
+  "main": "src/logger.js",
+  "types": "src/index.d.ts",
   "dependencies": {
     "fast-url-parser": "^1.1.3",
     "get-caller-file": "^2.0.5",
@@ -28,9 +28,9 @@
   },
   "scripts": {
     "benchmark": "bash ./scripts/benchmark-all",
-    "test": "standard && tap test/*.js && npm run test-types",
+    "test": "standard && tap tests/*.js && npm run test-types",
     "fix": "standard --fix",
-    "test-types": "tsc && tsd && ts-node *.test-d.ts "
+    "test-types": "tsc && tsd && ts-node tests/*.test-d.ts "
   },
   "author": "David Mark Clements",
   "contributors": [
