+++ tests/test.js
@@ -100,7 +100,7 @@ test('add transport.caller information when missing', function (t) {
 
   const logger = pinoHttp(options)
   logger.logger.info('hello world')
-  t.equal(options.transport.caller, join(__dirname, '../logger.js'), 'caller is set')
+  t.equal(options.transport.caller, join(__dirname, '../src/logger.js'), 'caller is set')
   t.end()
 })
